Move common props next to connection
