+++ src/cow-react/modules/wallet/web3-react/connection/walletConnect.tsx
@@ -2,7 +2,6 @@ import { ConnectionType } from '@cow/modules/wallet'
 import { getConnectionName } from '@cow/modules/wallet/api/utils/connection'
 import { useIsActiveWallet } from 'hooks/useIsActiveWallet'
 import { ConnectWalletOption } from '@cow/modules/wallet/api/pure/ConnectWalletOption'
-import { walletConnectOption } from '@cow/modules/wallet/api/pure/ConnectWalletOption/ConnectWalletOptions'
 import { TryActivation, onError } from '.'
 
 import { initializeConnector } from '@web3-react/core'
@@ -10,6 +9,13 @@ import { WalletConnect } from '@web3-react/walletconnect'
 
 import { RPC_URLS } from 'constants/networks'
 import { Web3ReactConnection } from '../types'
+import { default as WalletConnectImage } from '../../api/assets/walletConnectIcon.svg'
+
+export const walletConnectOption = {
+  color: '#4196FC',
+  icon: WalletConnectImage,
+  id: 'wallet-connect',
+}
 
 const [web3WalletConnect, web3WalletConnectHooks] = initializeConnector<WalletConnect>(
   (actions) =>
